requirements.txt
aws-cdk-lib>=2.150.0
constructs>=10.0.0
boto3>=1.34
aws-lambda-powertools>=2.36
pytest
moto
opensearch-py
requests-aws4auth